% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_imports_and_functions.R
\name{assert_dots_empty}
\alias{assert_dots_empty}
\title{Assert that no dots arguments are passed}
\usage{
assert_dots_empty(
  env = caller_env(),
  error = NULL,
  call = caller_env(),
  action = abort
)
}
\description{
This is an alias for `rlang::check_dots_empty()`, for
  consistency with other arguments. The function throws an error if any
  unnamed parameters were passed to the function where this is called.
}
\keyword{internal}
