% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_options.R
\name{assert_gp_options}
\alias{assert_gp_options}
\title{Verify that x is a valid \code{gp_options} object}
\usage{
assert_gp_options(x)
}
\arguments{
\item{x}{An object to verify.}
}
\value{
Unchanged input if valid, otherwise an error is thrown.
}
\description{
Verifies that \code{x} is a \code{gp_options} object (of class
\code{gitpins_gp_options}) and that all elements of the object are
valid for such an object. Use \code{gp_options()} to create \code{gp_options}
objects.
}
\keyword{internal}
