% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitpin.R
\name{clear_old_gitpins}
\alias{clear_old_gitpins}
\title{Clearing old gitpins is not currently implemented}
\usage{
clear_old_gitpins()
}
\description{
Clearing old gitpins is not currently implemented
}
\keyword{internal}
