% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitpin.R
\name{fstamp}
\alias{fstamp}
\title{Convert date-time variable to file-compatible timestamp}
\usage{
fstamp(the_datetime)
}
\arguments{
\item{the_datetime}{a date variable}
}
\value{
a character representation of \code{the_datetime} as
a timestamp with only file-compatible characters
}
\description{
Convert date-time variable to file-compatible timestamp
}
\keyword{internal}
