% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_gitpins-package.R
\docType{package}
\name{gitpins}
\alias{gitpins}
\alias{gitpins-package}
\title{gitpins: Pin urls to local file and version with git}
\description{
Quickly download an online resource and make it available
for R as a local file. Repeated downloads are versioned using
a local git repository, and if getting a document from the web
is not successful, a previous local download is made available.

To get started, check out the help for the \code{\link[=gitpin]{gitpin()}} function, or take a
look at the info on GitHub (see links below).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/torfason/gitpins}
  \item Report bugs at \url{https://github.com/torfason/gitpins/issues}
}

}
\keyword{internal}
