% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_dropper.R
\name{gp_dropper}
\alias{gp_dropper}
\title{Calculate refresh interval to target specific daily drop time}
\usage{
gp_dropper(
  drop_hour,
  drop_window = 1,
  drop_tz = Sys.timezone(),
  refresh_floor = 0.1,
  refresh_ceiling = 12,
  now_hour = Sys.time()
)
}
\arguments{
\item{drop_hour}{The hour at which the resource typically drops. Either
specified as a numeric in 24h time, or as a POSIX time object.}

\item{drop_window}{The window, in hours, during the time we expect the drop
to happen.}

\item{drop_tz}{The time zone of the resource. In particular, this is relevant
if the \code{drop_hour} is specified as a number, but \code{now_hour} is specified as
a POSIX time object.}

\item{refresh_floor}{The lowest allowed refresh interval.}

\item{refresh_ceiling}{The highest allowed refresh interval.}

\item{now_hour}{The hour at which the pin is taking place. Can be passed as a
numeric in 24h time, or as a POSIX time. Defaults to current time, which is
often appropriate.}
}
\description{
Assuming that one is looking to pin an online resource that is typically
updated at the same time each day, this function calculates a refresh
interval that maximizes the likelihood that a pin is refreshed close to the
time the new version "drops" (the \code{drop_hour}).

The assumption is that the resource drops sometime between \code{drop_hour} and
\code{drop_hour + drop_window}. During that time, the refresh interval is lowered
to \code{refresh_floor}. After the drop window closes, the interval slowly rises
to \code{refresh_ceiling}, in a way that ensures that a refresh is always
triggered if a resource has not been refreshed after the last drop window
closed.
}
\examples{
# Assuming that the resource is updated at noon, US/Eastern time zone
url <- "https://vincentarelbundock.github.io/Rdatasets/csv/openintro/country_iso.csv"
pin(url, refresh_hours = gp_dropper(drop_hour = 12, drop_tz = "US/Eastern"))

# Demonstration of the actual refresh interval returned by gp_dropper()
gp_dropper(drop_hour = 12, drop_tz = "US/Eastern")
}
