% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitpin.R
\name{gp_init}
\alias{gp_init}
\title{Initialize gitpins repository}
\usage{
gp_init(..., options = gp_options())
}
\arguments{
\item{...}{Reserved. All arguments must be named.}

\item{options}{A \code{gp_options()} object, used in particular to select
the directory for storing the pins (defaults to \code{here::here("gitpins")}).}
}
\value{
The path to the repository
}
\description{
This function is generally called automatically as needed. Note, however,
that to set a non-standard directory for the pinned files, this function must
be called before any other functions.
}
