% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_options.R
\name{gp_options}
\alias{gp_options}
\title{Create a gp_options Object}
\usage{
gp_options(
  ...,
  pin_directory = getOption("gitpins.pin_directory", here::here("gitpins"))
)
}
\arguments{
\item{...}{Reserved. All arguments must be named.}

\item{pin_directory}{A character string specifying the directory to use
for the pin repo}
}
\value{
A list representing the `gp_options` object.
}
\description{
Constructs a `gp_options` object with configurable parameters.
  The parameters can be set from the defaults, from R options, or directly
  when the options object is created.
}
