% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitpin.R
\name{init_gitpins}
\alias{init_gitpins}
\title{Initialize gitpins repo}
\usage{
init_gitpins()
}
\value{
The path to the repo
}
\description{
This function is called automatically as needed and should not
need to be called by the user
}
\keyword{internal}
