% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitpin.R
\name{list_gitpins}
\alias{list_gitpins}
\title{List available gitpins}
\usage{
list_gitpins(history = FALSE)
}
\arguments{
\item{history}{Should full (git) history be returned?}
}
\value{
A \code{data.frame} with the timestamps and urls of available gitpins.
}
\description{
List available gitpins
}
