% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitpin.R
\name{list_pins}
\alias{list_pins}
\title{List available pins}
\usage{
list_pins(history = FALSE)
}
\arguments{
\item{history}{Should full (git) history be returned?}
}
\value{
A \code{data.frame} with the timestamps and urls of available pins.
}
\description{
List available pins
}
