% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_options.R
\name{assert_gp_options}
\alias{assert_gp_options}
\title{Verify that x is a valid gp_options object}
\usage{
assert_gp_options(x)
}
\arguments{
\item{x}{An object to verify}
}
\value{
Unchanged input if valid, otherwise an error is thrown.
}
\description{
Verifies that `x` is a `gp_options` object (of class
  `gitpins_gp_options`) and that all elements of the object are
  valid for such an object. Use `gp_options()` to create `gp_options`
  objects.
}
\keyword{internal}
