% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_gitpins-package.R
\docType{package}
\name{gitpins}
\alias{gitpins-package}
\alias{gitpins}
\title{gitpins: Pin urls to local file and version with git}
\description{
Pin URLs to local file and version the pins with git.
Repeated downloads are versioned using a local git repository,
and if getting a document from the web is not successful, a
previous local download is made available.

To get started, check out the help for the \code{\link[=pin]{pin()}} function, or take a
look at the info on GitHub (see links below).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/torfason/gitpins}
  \item \url{https://torfason.github.io/gitpins/}
  \item Report bugs at \url{https://github.com/torfason/gitpins/issues}
}

}
\keyword{internal}
