% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitpin.R
\name{init_gitpins}
\alias{init_gitpins}
\title{Initialize gitpins repository}
\usage{
init_gitpins(..., options = gp_options())
}
\arguments{
\item{options}{A \code{gp_options()} object, used in particular to select
the directory for storing the pins (defaults to \code{here::here("gitpins")}).}
}
\value{
The path to the repository
}
\description{
This function is called automatically as needed and should not
need to be called by the user. Note, however that to set a non-standard
directory for the pinned files, this function must be called before
any other functions.
}
\keyword{internal}
