% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_pin.R
\name{tstamp}
\alias{tstamp}
\title{Convert date-time variable to sub-millisecond timestamp}
\usage{
tstamp(the_datetime)
}
\arguments{
\item{the_datetime}{a date variable}
}
\value{
a character representation of \code{the_datetime} as a sub-millisecond
timestamp
}
\description{
Convert date-time variable to sub-millisecond timestamp
}
\keyword{internal}
